Painter's Partition Problem
Given an array/list of length ‘n’, where the array/list represents the boards and each element of the given array/list represents the length of
each board. Some ‘k’ numbers of painters are available to paint these
boards. Consider that
each unit of a board
takes 1 unit of time to
paint.
You are supposed to
return the area of the
minimum time to get this
job done of painting all
the ‘n’ boards under a
constraint that any
painter will only paint the
continuous sections of
boards.
Example :
Input: arr = [2,
1, 5, 6, 2, 3],
k = 2
Output: 11

on Coding Ninjas : https://www.codingninjas.com/studio/problems/painter-s-partition-problem_1089557