Selection Sort
Selection sort is one of
the sorting algorithms
that works by repeatedly
finding the minimum
element from the
unsorted part of the
array and putting it at
the beginning of the
unsorted region of the
array.
You are given an
unsorted array consisting
of N non-negative
integers. Your task is to
sort the array in nondecreasing order using
the Selection Sort
algorithm.

coding ninjas Question Link:
https://www.codingninjas.com/studio/problems/selection-sort_981162